"10:40am. This is actually earlier than I thought I would wake up given how late to I went to bed yesterday. Let me skip the morning session. Let me chill, do the chores, have breakfast and then I'll finally deal with the modeling phase.
12:50pm. Let me finally start the modeling adventure. I'll get that rack done promptly.

1:05pm. I have that post. Now let me do the hooks.

1:30pm. The hook is turning out harder than I thought it would.

1:40pm. I asked in the general. It turns out that the sweep node produces intersecting geometry. The mesh it folding in on itself. This is quite annoying. What do I do about it? Rather than just waiting for an answer, I am going to have to do some research by myself. Maybe I should have done this in Blender after all.

1:55pm. I remembered the PolyDoctor node from the time I tried fixing the normals and that did the trick. It did not fix the self intersection though.

2pm. You know what, let me ignore this problem otherwise my whole day will be eaten away by it. Now that I've fixed the artefact it literally does not matter.

2:10pm. Actually, let me make a curve less vicious. I think I'd get this problem in Blender as well. Let me smooth it out a bit.

2:15pm. A wasted effort. I can't fix this.

Forget it. Let me move on. The next come pool improvements.

2:30pm. https://houdinihelp.ru/nodes/SOP/intersectionanalysis.html

Oh, wait, this actually looks for self intesections. And I had an idea, I could just blast the curve points that are near to the corner. I actually hand't even thought of this.

Ok, even if I have the intersecting points, I have no idea what to do with them.

2:45pm. Ok, I blasted the points around the sharp corners and then I beveled the edges. This is the solution. Sigh.

2:50pm. This wasted so much of my time. Let me focus on the pool.

Ok, I need the pool front group, and I need it before it has been extruded.

I should try the facet node.

Also some of the edge tools I should check out here. I am not happy with blasting the hook.

3:20pm. Holy shit, how I hate that the input geometry to copy to curves is the second input. Couldn't they have fixed this early on?

3:40pm. Ok, let me do the pool ladder. I did the nibs, but for all I know my proportions could be badly off. but it is fine. I am doing everything procedurally, so it will be easy to change things.

The ladder. Let me get to it.

5pm. Did the ladder, but after doing the booleans on the pool, now I notice that the ngons are messing up the lighting in unnatural ways. Let me deal with this.

https://youtu.be/eD-W9jh9hJE
Grouping Ngons (or Quads or Triangles) and Fixing them

Let me go with this. Yes, this, not the Cop Craft thread.

5:20pm. None of that stuff did jack. What did was settting up the normals so that they resemble flat shading.

5:25pm. Ok, that is enough of that. Doing it by face area works well enough. The issue is that after the boolean ops, the topology got messed up. Had I done this by hand, this would not have happened.

But it would have been annoying to do. Let me just move on.

The last thing that I need to deal with is the walkway. I did the HellProto logo, but I do not think it is suitable to put it there. I should be using the pool segment as a guide.

6:35pm. Had to leave for lunch. Let me have this fruit and then I'll do the walkway.

6:55pm. Let me resume. I want to figure out how to make the cross section for sweep. It is a real pain in the ass.

7:05pm. I've gotten the hang of how to do the cross section.

8:05pm. https://www.hipflask.how/geometry-essentials-03
> 9. Using the Carve Node to Shorten the Curve

I need to check out the carve node. The way I am currently makign the curves shorter is ridiculous.

8:10pm. Damn it, how do you get the length of a spline? If I could figure that out, I could get rid of all the uses of boolean curve SOP.

8:30pm. I figured it out. Measure outputs the perimeter. I used this before, but then completely forgot about it.

8:35pm. >  Warning  Local variable 'primnum' not found.

Enough. I'll figure out how to use primnum tomorrow. I have no idea why it is failing for me. I'll have to take a look at the carve examples first. Right now, I do not feel like it and keep looking for shortcuts.

Actually, let me just try if I can use the attribute directly.

Nope. Also I am not wrong. I found in the cookbook that @primnum is the right way to get the primitive local number.

8:50pm. Let me play Fallout. Forget geomentry nodes."