"11:25am. My sleep tonight has been bad. I must have been obsessing all night about my regrets. I really don't know what to do about them except make money. Money is ultimately the difference between an adult and a kid. A kid has nothing except what his parents give him. And that is exactly my situation currently. I need to make money. I can't keep scamming myself with ML. If it is really meant to be, maybe that social media post will give me a pull and I will be able to get my hands on the AI chips on schedule. Otherwise I'll have to proceed on my current path and hope it connects with the old one at some point.
11:30am. Well, let me chill and then I'll have breakfast.

12:30pm. I am having breakfast. I watched the 4th Solaris video and I had enough of that. The Solaris stuff is pretty much useless to me apart from allowing me to use Karma in the viewport. There is no reason for me to learn this, it is such a waste of time.

When it is time to start what I am going to do is get back to texturing. I need to do the pool water first.

2:05pm. I am back. Let me start.

Actually, before the pool water, I need to figure out how to load the textures into the MaterialX shader.

Before that, I want to work more on the pool UVs. I should separate the inner part from the rest properly, so it does not overlap...

No it does not matter. This is because it not visible. It is just my perfectionism talking.

Then, let me load the textures.

2:15pm. Huh, what was the Mtlx Principled Shader? Wasn't it Mtlx Surface? But that can't be it. Let me check out the video again.

https://youtu.be/tI4zqTezniI?t=452

It is standard surface.

2:25pm. I need to ask a question in the Houdini thread.

https://www.youtube.com/results?search_query=houdini+materialx

Also I should watch some of these tuts. I really have no idea how to load textures.

2:35pm. Asked. Focus me. Focus on loading the textures.

2:40pm. No, I can't figure it out. The nodes are simply not working. The roughness is not working for some reason either.

Why does the viewport not respond when I move the sliders?

2:45pm. Oh, it worked. What the hell, I only changed the contexts. This is Houdini being quite buggy.

What do I do about the rest of the textures? There is no way around it, let me watch the tutorials.

https://www.youtube.com/watch?v=FVmFFn7EPdA
Building Shaders With MaterialX in Solaris | Houdini

https://youtu.be/FVmFFn7EPdA?t=276

What is this Mtlx Image Albedo. He says the texture comes in here.

Yeah, it works.

And now the viewport is not reacting properly again.

Sigh. Also I can't see displacement on the Mtlx Surface.

Also I've completely forgotten what specular is supposed to be doing.

https://youtu.be/FVmFFn7EPdA?t=309

Here is something regarding displacement.

https://youtu.be/FVmFFn7EPdA?t=480

Let me pause this here. This is too advanced for me right now. I am not interested in how to build the principled shader from scratch.

https://youtu.be/QjDMZAJAeCs
Enabling Displacement in Material X and Karma | Houdini 19

Let me watch this first since it is only 5m.

https://youtu.be/QjDMZAJAeCs?t=153

There is a restart render option.

3:10pm. So far my impression that this is lower quality than Blender. Cycles is always responsive and does not require manual restarts.

https://www.youtube.com/watch?v=ABF7op2W8pM
Karma XPU and MaterialX | Adam Swaab | Houdini 19 HIVE

Let me watch this. I might end up dropping Houdini here after all.

3:30pm. Yeah, before I can progress, I need to figure out why stage is not importing normals. I really hate Houdini when it does things like this.

3:50pm. It turns out I forgot to add the normals to the object. But now that I've fixed that, I find that the UVs are messed up in Karma XPU.

Hmmm, it seems the UVs work fine with regular Karma.

4pm. https://youtu.be/ABF7op2W8pM?t=1675

Here the actual tutorial part starts.

Sigh, I wish I could be faster in figuring out how to get the rendered going. I really wanted to do the pool instead of fiddle with setting it up.

4:05pm. In Blender I want to try something out. I had the idea for it during the night and thought about it only now. Is there a select parent option?

Yep. You can access it with Shift + G. The brilliant thing about Blender is that it remembers the last use and positions the menu so the cursor comes under it. This UI design is a far cry from Houdini who always expects you to haul everything yourself.

4:25pm. I think I am fucked with Karma XPU.

Hmmmm...

Let me try changing things a bit. The overlapping UVs might be the problem.

4:45pm.

///

>>883907
>Is that object set to render as subdivision surfaces in /obj by any chance?
No. Maybe it has something to do with the way the object tree is structured. I have packed prims and another object that does not have UVs, but I do not want to start taking things apart to figure out what is wrong.

I think at this point I have two options:

* Do the lookdev in Houdini using the regular rendered and let the Karma CPU grind away at the finished scene for a while.
* Find a way to export the scene into Blender and do the lookdev there using Cycles.

Yesterday I watched the Entagma guys talking how easier it is to do the lookdev in Blender, and I am mostly interested in Houdini for generating procedural environments rather than doing simulations, so that kind of workflow might be a good fit for me. The reason I decided to try out Houdini is to generate those big vines.

What are my options for exporting given that I have the non-commercial license? The regular export options are grayed out, but I saw a video on exporting geometry into .obj using the File SOP. I've never tried it so I am not sure whether that would deal with packed instances correctly. There are about 20k of them in the scene and I do not want them to be put into a collection, let alone output as separate geometry.

///

Let me finish the tutorial. I'll just use the CPU renderer. I want to go into Blender, but that would lead to its own problems. I'll have to do the scene in Houdini. Using the raster renderer is good enough for my purpose.

4:50pm. I am trying to save the vines to file, and given how long it is taking most likely it is outputting everything.

Holy shit, where did it put the file?

I should have put `$HIP/` before the file name.

Now I have a 2gb obj file wasting space somewhere.

It is in the $JOB folder...Which is also my Users folder.

Deleted.

5:10pm. I am just wasting time here now. I do not need the tutorials, but raw grit instead.

It was a mistake doing this in Houdini after all. But I can finish the scene and have it look good regardless.

5:15pm. Hmmm, is the MaterialX shader not working with the raster rendered after all?

5:20pm. Forget the MaterialX shaders. I do not know why, but they stopped working with the regular view. I could have sworn it worked for me earlier.

This is just another aspect of Houdini's general shittiness at rendering.

I am going to have to use the Principal Shader. I guess Mtlx stuff is in alpha as well.

5:25pm. What the hell is going on here? Why did the base texture suddenly stop working.

5:35pm. I have no idea, I had to create a new shader.

6:40pm. I am starting to hear thunder let me close here early."