"11:25am. I slept well, but decided to stay in bed for a while. I've had to time to think about what I want to do, and I'll think I'll watch the cloud tutorials by Rohan. I should play around with volumes for a bit.
11:45am. Let me have breakfast here. I want to start, but this is no time to be doing that.

12:20pm. Another chap and then it is chores time.

12:45pm. Let me do the chores here. I've done enough morning slacking. There simply aren't enough hours in the day to satisfy me.

3:05pm. I am back. Let me chill a bit first. I need to cool down.

3:40pm. I've cooled down. It if far later than usual, but let me finally start.

It is just so hard to find the time to do this even though I have the whole day to myself.

Forget the tutorial by Rohan. Let me check out some of the stuff on Youtube.

https://youtu.be/WTg_pVZoac0
Free Masterclass: Developing Complex Details Using Volumes in Houdini

Oh, this is 3.5 hours. This will teach me a lot about volumes.

https://youtu.be/WTg_pVZoac0?t=19

Looking at the TOC, he covers walls and bricks. This is not what I imagined volumes would be for. I'll definitely benefit a lot from this. These cliffs look amazing.

https://www.youtube.com/c/Rebelway/search?query=free%20masterclass

This channel has hours long courses. What a find. I wonder why the volume one does not show up in search? Ah, it does. It is the second one.

https://www.youtube.com/playlist?list=PLrwHqtPQxnbY7kjiqnTNhvvQFeX39rMPD

Here is the playlist of them. It would take me two weeks at least to go through all of these.

https://youtu.be/WTg_pVZoac0?t=251

He mentions that VDB is better than Houdini's native volume.

https://youtu.be/WTg_pVZoac0?t=327

Let me follow along. The video tutorials are much better than reading docs.

4:20pm. https://youtu.be/WTg_pVZoac0?t=546

This is so confusing. I am playing with volumes and am noticing that its rendering is completely different depending on the absolute size of the sphere. Setting the voxel size does not adjust for that.

https://youtu.be/WTg_pVZoac0?t=625

I do not get why the noise pattern looks so much better for him. When I try it, I get areas that are extremely light and dark.

Let me ask.

///

>>884062 (You)
Try hscript point() expression

///

Hmmm, this is for that attribute question. Let me save what I am currently doing and I will try it out. If I can use the @ptnum inside the expression it could work.

5:40pm. Let me stop here for lunch.

6:35pm. What the hell? How did it last this long. At any rate, no way I am going to stop here. I am going to keep going till 10pm.

7:25pm. https://youtu.be/WTg_pVZoac0?t=1914

The voronois fracture changed significantly.

8:10pm. AHhhhhhhhh. The Toogle visualization is that small icon after piece prefix.

8:35pm. Holy shit, I just can't figure out what the debug flag for VOPs does. Right now I want to look into the random node to see why it is not outputting the y axis and can't figure out how to do it.

8:40pm. Maybe it is just an support flag without any explicit effect.

///

>>884262 (You)
the ramp parameter node?
there's a weird quirk to it - you can't tweak it from inside the vop. go up a level and you'll see the interface exposed on the vop parameters. by default it's set to zero. don't ask me why.

///

I did not expect this. I'll have to keep this in mind for the future.

8:50pm. https://youtu.be/WTg_pVZoac0?t=2375

It has taken me a while to remember that moving the offset is not for the points, but for the noise itself. What he is doing here is spew, he should have just turned on 3d output.

Let make sure. I'll try outputting the thing as an attribute.

Bind Export VOP is what I need. And I see that the 3d outputs are in fact random. Great. This is much better than the solution in the video.

Hmmm, for storing the distance of each piece, I should be using a dictionary rather than attaching it to primitives. How do I do that.

https://www.sidefx.com/docs/houdini/vex/dicts.html
> The dict type will be considerably less efficient than expanding the dictionary as a variable per entry; so should only be used where the set of variables and types are varying.

I was wrong. I guess I should just attach the dist value to primitives.

9:20pm. I've played around quite a bit. I did managed to extrude every individual fractured piece.

https://youtu.be/WTg_pVZoac0?t=2552

Did this guy forget that each piece already had a name attribute?

But at least this is educational.

https://youtu.be/WTg_pVZoac0?t=2632

He is packing to get the centroid. Wasn't there an extract centroid. He also could have iterated over each individual piece and got the center of the bounding box. Also, I think I saw the extract centroid node while looking around.

Extract Centroid SOP, just as I suspected.

9:30pm. I could use that. But the Voronoi Fracture in fact does give the centroid information as its second output.

The approach this guy is taking can be unoptimized. This demonstrates the danger of just following courses blindly.

I am making slow progress, but this way of going through it is pretty immersive.

Now what is he doing what that rest attribute?

> Creates rest attributes on the geometry so procedural textures stay put on the surface as it deforms.

This is possible?

https://youtu.be/WTg_pVZoac0?t=2731

9:50pm. point("../centroids",detail("../foreach_begin2_metadata1/","iteration",0),"P",0)

It is a huge pain in the ass to access something in hexpressions

10:30pm. Voronoi fracture has an option to output the primitive piece in addition to the connectivity.

10:45pm. I am still at it, trying to figure out optimal ways to implement what he is saying. My own methods are actually better than what is show in the vid, but on the flipside it takes me forever to do it.

https://youtu.be/WTg_pVZoac0?t=2731

Let me watch this just for a bit and then I will stop.

10:50pm. Let me stop here. I actually went quite far with this today. I was actually considering stopping a lot earlier despite what I said.

I figured out a cool trick on how to import attributes as detail attributes in a loop so you don't have to use channels in Hexpressions. You just use the various import nodes and re-export them. This is annoying to do, but it a lot better than having to mess around with copy pasting channels and whatnot.

I am going to have to redo the rotation to match what he is doing. He is using the AA noise inside a VOP node while I am doing something else. Is AA noise the simplex noise?

...No, no it is not. I can only use AA noise in VOP nodes unfortunately.

11am. Tomorrow, I am going to put a dent in this. I've gotten sidetracked from my original goal, but learning about volumes is a worthy endeavor at this point. I really am curious how his did those super realistic cliff, wall and rock faces."