"2:15pm. https://www.patreon.com/posts/50498890 Heroine Survival Chapter 1
I've been meaning to pick this up for a while as I have a soft spot for killer girls. But I just wanted to note that this is posted on Patreon of all things. I didn't know that this was possibel. I'd of course want to make my own place as soon as possible, but until I get some kind of income will post stuff on places like Twitter and potentially here.

Let me leave this for later. Let me finish the Mantra shading tutorial and then I will get started on my own stuff. I need to master shading, and it won't be too hard.

https://youtu.be/wXf6PUdGZtI?list=PLFkMNnEYa3APIjPiwdEFL6BOJWLLqk7_S
Types of Lights, Global Illumination, Render Noise and Settings — Rendering with Mantra ep. 10

Let me start this.

2:35pm. https://youtu.be/oUWFxqhJDlA?list=PLFkMNnEYa3APIjPiwdEFL6BOJWLLqk7_S
Car Paint Shader, Fresnel, Frosted Glass Effect — Rendering with Mantra ep. 11

Let me start skimming. I do not feel like watching all of this.

https://youtu.be/oUWFxqhJDlA?list=PLFkMNnEYa3APIjPiwdEFL6BOJWLLqk7_S&t=907

He says that the sheen makes normals which look outwards the camera look lighter.

I guess I have an answer what that is.

3:05pm. I am dead, I can't go on anymore with these tutorials.

https://docs.chaos.com/display/VRAYHOUDINI/V-Ray+Material

Right now I am reading the docs for a bit. I was wonering what Coat is, but I think it is literally another layer on top of the first one.

3:35pm. No, that is not true. The coat literally just multiplies the numbers of the diffuse color. I was confused as to how the changes are going, but it is a straightforward multiplication. This can lead to unexpected results unless you use a grayscale value for the coat.

I see why Blender calls it clear coat. I could use this as a mask.

3:40pm. But it not a huge deal. Let me check out sheen.

3:50pm. A little bit of experience goes a long way. Sheen is very similar to frenel. The main difference is that the diffuse color gets interpolated with the sheen color. I did not understand this while I was back at Blender.

4:20pm. Let me take a break. I think I will step away from the screen for just a bit.

I've been going at it tirelessly for over a week, but now it is time to step back just a little bit and review.

I...can do it.

5:20pm. Let me resume. I used to do these motivational brainstorming sessions all the time when I was programming.

Let me just play with mirrors a bit and then I will go forward with my plan.

5:40pm. Hmmm, I see.

When I was messing with Blender I did not really think too deeply about the index of refraction. But now I see that if you crank it up, it is possible to make dielectric surfaces into mirror.

My initial impression was that dielectrics have weaker reflections, but now I understand that the situation is more nuanced. For dielectrics their reflectivity is strongly dependent on the IOR. Back then I did not want to mess with the parameter, but now that I've done so I feel that I've learned a lot. IOR of water is 1.3, right?

https://knowledge.autodesk.com/support/3ds-max/learn-explore/caas/CloudHelp/cloudhelp/2015/ENU/3DSMax/files/GUID-CCD9B76C-9AC6-46E6-8B9C-E367CFC0FDAF-htm.html

Hmmm, I see. Even air itself is 1.0003.

6pm. I needed to do this bit of thinking.

8:10pm. I am still thinking about it. I've been doing a skill check in my head, and have been stuck on figuring out the best way to do a pattern that I once saw. It was a radial block pattern.

8:20pm. This issue I am grinding away at right now is so annoying and it does not matter in the least. I am being dumb by even thinking about it.

At any rate, going over all that I've learned in the last...7 weeks, I can tell it is quite a lot. I know a lot of modeling, and in the last week I learned quite a bit about shading.

8:45pm. Ok, I figured the problem of how to move the centroid out from a poligon. Forget this for now.

To reusme the earlier thread, I've learned an enormous amount of Houdini at this point. What am I still missing?

Well, I go back to the start and watch that spiral tower video. There are also video on the Houdini site.

But that can be put aside for now. I certainly know enough.

Yeah, I am done with the tutorials for the time being.

It is time to put what I know into action. I've learned the equivalent of sketching in Houdini and now it is time to move to painting.

I've gone over all that I can with geometry, and all that I can do with shaders. It might seem fancy, but what it allows me is to be smart with repetition. It might be a definite step up in abstraction compared to Blender, but it is still 3d art at its root. All it does is allow me to add a little programming flair.

9:10pm. It is going to be good when I get going.

Honestly, I've been kind of slow, and I do not know how much that is going to improve. I've never been a speed demon. In that respect, going for a VN first is a not a bad idea. If I went for a manga, I'd probably burn out.

9:15pm. Let me close here 'early' for a change. Tomorrow, I am definitely going to start grinding things out. It really is not that complicated. Put the diffise, put the falloff, put in the translucent, put in self illumination, add some noise on top of that. The vines are not something that should give me too much trouble. I just have to let go of my pretensions.

It is geometry itself that shoudl be the hard part of doing art. By the time you are done with sketching yuo should have the scene fully visualized.

9:15pm. One thing I regret is the way I did the flower. The way I've done it would make some things more difficult. Instead of doing a sweep, I should have made flower and then either bent it or path deformed it. That would have allowed me to have stable uvs, for one. I could use that for some special patterns on the vines.

I think I am going to go for that. There is no reason I can't redo the core parts of the scene. In Blender this would be hard, but in Houdini it is a piece of cake.

But I should not get fancy right at the start. Let me do the basic thing and then I'll consider what to do from here. I would not be surprised if the basic thing turns out to be good enough.

The important thing is to just sit down and just get things out.

Those AI chips are waiting for me and Spiral is sitting there waiting to be used.

I have two choices:

* Get a job and get the resources that I need to chase the Singularity that way.
* Acquire manipulation tree skills and target the world directly.

Either I can sell myself into slavery and join the NPC ranks, or I can take risks. I did get close to getting a job that one time. If the offer was not 2.5k, but something reasonable, I'd have taken it and by now have enough money to afford the chips and a lot of other things besides. But it would be a pity to withhold Heaven's Key from the world. I'll make it a blow it won't recover from. I'll get a good hit, get the money for the chip by the end of it and acquire the power I've been dreaming about.

No matter which path, capitalism will take its bite out of me.

A scholar sells his intellect.
A laborer sells his strength.
An artist his dream.

But rather than be an artist, I should be a prophet. Rather than selling my brain, muscle or heart, I should sell my eyes.

Dreams come and go, but my vision will be neverending. That is what I should aspire to.

Instead of relying on my brain, I should rely on my desire. My will, Heaven's Key will happen."