"9:50am. I am up. Let me chill just a bit and then I will start. I am going to watch those short procedural texturing vids on the top of Youtube and then dive into the 5h course.
10:15am. https://youtu.be/9ngbOV0YKbk?t=240
Advanced Procedural Texturing

Let me resume this from yesterday. This is using some program I've never heard about called Clarisse.

https://youtu.be/j-B5Cib_DXw
Displace Geometry in Houdini || Quick Tricks

Let me watch this first, I actually don't know how to apply displacement on geometry directly.

https://youtu.be/j-B5Cib_DXw?t=34

Ah, displace is actually in a VOP.

10:30am. Let me go back that other tutorial. I've played with displacement a bit.

https://youtu.be/oCQhnVl-xrI
What is Clarisse?

10:35am. https://youtu.be/9ngbOV0YKbk?t=424

I've been playing around with UV noise in the V-Ray shader nodes, but it never worked for me for some reason.

https://www.reddit.com/r/vfx/comments/lsxu3n/what_is_clarisse_used_for/
> Yeah as already mentioned, it can handle a huge amount of polys and have them viewable in the viewport instead of bounding boxes. I am talking billions. Its definitely suited towards environments.

> I would say if your really technical and perhaps know some Houdini then continue down that path. But its really good for people coming from a more DMP and creative background. Its quite easy to get scattering and make environments quickly. I think you can get creating stuff quicker than you would if learning Houdini from scratch.

> I actually made a timelapse the other day of a project I made using it, will give you some idea of it: https://www.youtube.com/watch?v=Fo7pOE5_2EU

> You can see by the end that the scene had 480 million polys in it just fine.

This is interesting.

> "Crashisse" - wish we had thought of that. It was painful. Open a scene, wait 40 minutes to load, make a few changes, crash, repeat. In a normal working day, we got about an hour or two worth of work done.

What the hell? Apparently this was an issue in old versions.

> I absolutely love clarisse for env work and there is nothing like it out there where you can make changes and experiment like you can in clarisse. Houdini is an amazing software that I love and use daily but it is not a clarisse alternative. Houdini can't display 1/1000000th the geometry you can display and move around in clarisse so you are looking at proxies of proxies while working on big env's as opposed to being able to like "hey let me move around this 6 million poly tree in realtime in render view while it's rendering to find the best composition" like you can in Clarisse.

10:50am. https://youtu.be/9ngbOV0YKbk?t=840

I can't really follow this, but it does not matter.

10:55am. https://youtu.be/9ngbOV0YKbk?t=1038

I saw this in Blender. What is a point cloud texture?

Ohhh, this is what I had in mind for doing the water around the pool. I should look up point clouds in Houdini.

https://youtu.be/RfaBNoicXp4
Find Points on Geometry with Point Clouds | Houdini VEX Quickies

Let me do a bit of research in this. This might be an alternative to painting masks. Instead painting masks, I could drop down points and transfer attribute. That would allow the system to be persistent to the changes in the underlying geometry. It actually has a lot of advantages compared to mask painting.

https://youtu.be/uTPR0cUVRdg
Magic Market | Procedural Rocks 4 | Procedural Texturing

I'll skip the Blender tutorials as my focus on is on Houdini. Let me watch this for a bit. After that I'll dive into the 5h tutorial from the start.

https://www.sidefx.com/tutorials/magic-market-procedural-rocks/

I'll leave this for later if at all. Let me just go with the Rebelway tutorial. I do not need so many rock tutorials.

11:15am. https://docs.chaos.com/display/VMAX/VRayDistanceTex

I just remembered I saw the distance texture. Maybe this is what I need for those water puddles around the pool.

Also I just realized - in the material palette I have the assets from the V-Ray library. This will make using them a lot easier.

https://docs.chaos.com/display/VRAYHOUDINI/V-Ray+Distance

This could be really useful for water. I can just place a bunch of points and pass this as the point cloud. Brilliant. This resolves one of my issues.

I just remembered that Houdini has the Spray Paint node. I could use that to spray a path.

11:25am. This is wonderful. If you deform the geometry in the earlier nodes the sprayed points get projected onto its surface. It has a lot of roboustness. It seems the projection takes into account the original orientation of the points rather than just placing them on the nearest point. Nice. Though it might be better if it projected the stoke itself rather than the points. It does not matter right now.

I am satisfied. Let me go back to the master class.

12pm. https://youtu.be/kz-_z3wREbY?t=1658

Let me skip this pipeline stuff. It is not something I need to be concerned about. Let me go back where I was yesterday.

https://youtu.be/kz-_z3wREbY?t=10611

Oh, this thing he does to put in regions is brilliant. Why did I never think of this? I should keep it in mind.

12:40pm. https://youtu.be/kz-_z3wREbY?t=11524

Let me stop here in the middle of part 2.

This is good enough.

Let me do the chores and have breakfast here and I am going to try going with the mask painting plan. It should work well for the leaf.

I've got this thing in the bag. It won't give me trouble for much longer. I can do an enormous amount without ever touching UV unwrapping by hand. And if I use UV projections not even textures themselves will give me particular trouble.

I can just tream them as a different kind of noise. For specific things there will always be masks and geometry.

I am thinking back to that banana tutorial, and with the understanding I have now I'd have no trouble dealing with it without having to go into an app like 3d coat.

12:50pm. Let me do the chores."