"2:35pm. https://nautil.us/deep-learning-is-hitting-a-wall-14467/
If symbol manipulation is to be merged with NNs, as Marcus recommends, then better hardware is needed. GPUs aren't a good fit for that approach. AI chips could do it.

I wanted to make that comment. Let me finish reading this and I will start.

2:40pm. Let me start.

It is time to go into the action. Let me take care of that leaf.

2:45pm. Oh wow, the attribute paint even considers stroke. It is very nice!

2:55pm. Hmmm, instead of putting the material outside, wouldn't it be possible to put it inside as a detail attribute? That would be ideal.

3:05pm. Ah, I've just realized that the material assignment must always be a primitive. So no detail. But it does seem like it would work on individual faces.

Oh that does work. Interesting. This has potential.

3:10pm. https://www.sidefx.com/docs/houdini/nodes/sop/pack.html

I really want detail attributes to go from and to the packed primitive.

3:25pm. Forget trying to preserve the old architecture. For things like subdividing the petals and the leaves doing that before packing them will work fine. I can have two nulls as two different resolution outputs. Let me do the branches just a tad.

3:40pm. Why are point attributes being interpolated in such a strange manner?

4:05pm. Ok, at least the subdiv interpolation looks great.

The way they are interpolated in the viewport and in V-Ray is different, but it is not a problem ultimately.

4:15pm. I am thinking way too hard about this as usual. Let me just move on.

Wow, I hate that I can't just unpack, subdiv and repack like before. Let me ask about this. It is an real eyesore.

You know what, when unpacking I'll just transfer them to primitives and back. That is the easiest way. Just why am I racking my brain over this?

5pm. Interesting...actually let me test this with instancing before I just to conclussions.

Yeah, great. It does not work on faces, but when nesting packed primitives, it is fine if the outer primitives do not have material paths. It will just get them from the individual ones. Also I just realized why transfering the detail attributes is not done. It would lead to havoc in situations such as these where I could be packing an unpacking many different things.

5:45pm. It crashed. Damn, I should have been saving more often. I got a sense I am not doing it enough. I tried to do a quick save before the app closed in the middle of a crash and it corrupted the file. Thankfully Houd often makes backups.

5:50pm. I am a fool. Houdini is not like Blender. I should have known that trying to change structure of the nodes while it is rendering will lead to it crashing. I should have stopped and resumed the rendering. And saved before doing that.

The 'operations' nodes could not be designed any more annoying. It is hard to believe this is version 5. The plugin has been on the market for this long and they still haven't fixed this broken design. Once I have the money I am tempted to switch to Redshift just for this.

6:30pm. It looks like a funky leaf. Let me make a separate mask which is inverted for the bottom side.

8:10pm. How is it this late? I barely did anything despite going at it for the entire day.

8:35pm. Fuck V-Ray! It seems the floating point operation power node does not work. For fuck's sake!

8:45pm. I actually don't need the power op that much. The same role can be taken up by remap. I put in some sparse noise and took the min with respect to the original hues. It works well.

Once again, I've spent an entire day what should realistically take me 10m to do. But I did learn a bit.

It is actually pretty difficult to work like this. It really makes a huge difference whether I have real time feedback or 10s feedback. It makes a difference how easy the nodes are too use. It might have been better to use textures after all.

8:55pm. I've turned on the frame buffer a while ago, but I do not see any option how to save the existing image. I'll have to figure that out. For some reason the falloff texture is not really working as expected.

9:05pm. I did something pretty good now. It turns out you can just pass in a color to self illumination. now instead of white sphere, I have a yelow gradient at the borders. I made use of the falloff texture like this.

Ah yes, I might have painted the petal, but I lost that work due to the crash.

9:10pm. Note to self, use the attribute remap to put the mask in the [0,1] range. It would have made things a easier in the shader part.

9:15pm. Maybe I should have used Distance Along Geometry even for the leaf.

Right, I could have used groups to factor out the inner parts. That would have worked well. But it is not like that I did was hard. Still, I should keep it in mind. I'll probably be doing mask painting a lot in the future. I am kwetching about how textures would have been better, but I should see my current approach to its conclussion.

9:40pm. https://docs.chaos.com/display/CWVRAYMAX/GPU+Dispersion

What is dispersion?

9:45pm. No, I won't use this for the petals.

9:55pm. Let me work on the stalk.

10:10pm. Damn it, the power really does not work. I will literally have to use the remap for everything.

10:20pm. Let me work on the roughness settings for the stalk.

To start with just how did this work in Blender? V-Ray has a reflective color which I find confusing.

11:35pm. Let me just put some noise on the stalk and I will call it a day. The fatigue is really catching up to me.

11:35pm. The UV noise options really don't do a thing for some reason.

11:45pm. Random crash. Thankfully I've been saving regularly.

11:50pm. It crashed again just as I turned the bypass of on a remap node. The people working on V-Ray really aren't into concurrency or reactive programming. Or memory management.

12:10am. I am done with the flower. If I was actually good at this, I could have done it in half an hour or less. But I did get some insights today. While picking colors, I realized that the contrast between 0.01
and 0.03 blues is just as visible as between 0.1 and 0.3, and between 0.3 and 0.9. Light intensity is on a log scale up to a certain range. I think now that I understand this, my color picking should improve.

3/11/2022

9:05am. I am up. Did I forget to commit yesterday again? Yes I did. At any rate, I slept quite well. I do not know how well suited for art I am, but I am certainly motivated to do it. I am going to get started right away.

Today I want to find out two things:

* Why the UV noise in V-Ray does not work.

I thought that I was doing something wrong, but given the extreme betaness of its Houdini intergration, for all I know it might be broken.

* Try out Redshift.

Redshift does not have a crack, but I had the idea of simply rolling forward the trials using fake emails. I am going to try that. Let me state this: if it was not for V-Ray's buginess I would not even be considering it right now. I meant to get a V-Ray subscription when the the time was ripe, but now I am not sure I want it. Let me get on with it. Today I am going to stop for a while after lunch. Yesterday the session was so long that I missed the timming to take a bath."